requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
openai
python-dotenv
pymupdf

torch
torchvision
numpy
Pillow

scikit-learn
matplotlib
seaborn
datasets

